third_party/blink/web_tests/external/wpt/html/rendering/replaced-elements/attributes-for-embedded-content-and-images/video-aspect-ratio-expected.txt

This is a testharness.js-based test.
[FAIL] Aspect ratio for regular video
  assert_approx_equals: aspect ratio before load expected 2.5 +/- 0.001 but got 2
[FAIL] Aspect ratio for regular video with width:auto
  assert_approx_equals: aspect ratio before load expected 2.5 +/- 0.001 but got 2
Harness: the test ran to completion.
